Changed to 8 Decimal places
Changed payment->amount to 8 decimals
Changed utxo->value to 8 decimals
added more info for psql on CreateUTXO

+++ pkg/store/sqlite.go
@@ -59,7 +59,7 @@ CREATE TABLE IF NOT EXISTS payment (
 	id INTEGER PRIMARY KEY AUTOINCREMENT,
 	account_address TEXT NOT NULL,
 	pay_to TEXT NOT NULL,
-	amount INTEGER NOT NULL,
+	amount NUMERIC(18,8) NOT NULL,
 	created DATETIME NOT NULL,
 	paid_txid TEXT,
 	paid_height INTEGER,
@@ -74,7 +74,7 @@ CREATE INDEX IF NOT EXISTS payment_paid_height_i ON payment (paid_height);
 CREATE TABLE IF NOT EXISTS utxo (
 	txn_id TEXT NOT NULL,
 	vout INTEGER NOT NULL,
-	value NUMERIC NOT NULL,
+	value NUMERIC(18,8) NOT NULL,
 	script TEXT NOT NULL,
 	script_type TEXT NOT NULL,
 	script_address TEXT NOT NULL,
@@ -651,7 +651,8 @@ func (t SQLiteStoreTransaction) UpdateChainState(state giga.ChainState, writeRoo
 func (t SQLiteStoreTransaction) CreateUTXO(utxo giga.UTXO) error {
 	// Create a new Unspent Transaction Output in the database.
 	// Updates Account 'incoming' to indicate unconfirmed funds.
-	// psql: "ON CONFLICT ON CONSTRAINT utxo_pkey DO UPDATE ..."
+	// For psql: "ON CONFLICT ON CONSTRAINT utxo_pkey DO UPDATE ..."
+	// Remove in the end "WHERE txn_id=$1 AND vout=$2"
 	_, err := t.tx.Exec(
 		"INSERT INTO utxo (txn_id, vout, value, script, script_type, script_address, account_address, key_index, is_internal, added_height) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10) ON CONFLICT DO UPDATE SET value=$3, script=$4, script_type=$5, script_address=$6, account_address=$7, key_index=$8, is_internal=$9, added_height=$10 WHERE txn_id=$1 AND vout=$2",
 		utxo.TxID, utxo.VOut, utxo.Value, utxo.ScriptHex, utxo.ScriptType, utxo.ScriptAddress, utxo.AccountID, utxo.KeyIndex, utxo.IsInternal, utxo.BlockHeight,
